+++ scraper/src/main/java/GroceryFamily/GroceryDad/GroceryDad.java
@@ -7,47 +7,27 @@ import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-import java.util.concurrent.Executors;
-
-import static java.lang.String.format;
 import static org.openqa.selenium.chrome.ChromeDriverService.CHROME_DRIVER_SILENT_OUTPUT_PROPERTY;
 
 @Slf4j
 @SpringBootApplication
 class GroceryDad implements CommandLineRunner {
-    private final GroceryDadConfig config;
+    private final ProductAPIClient client;
+    private final Scraper scraper;
 
-    GroceryDad(GroceryDadConfig config) {
-        this.config = config;
+    GroceryDad(ProductAPIClient client, Scraper scraper) {
+        this.client = client;
+        this.scraper = scraper;
     }
 
     @Override
     public void run(String... args) {
-        var client = new ProductAPIClient(config.api.uri);
-        //noinspection resource
-        var threadPool = Executors.newCachedThreadPool();
-        try {
-            for (var name : config.enabled) {
-                threadPool.execute(() -> {
-                    Thread.currentThread().setName(name + "-worker");
-                    var visited = Scraper.scrap(scraperConfig(name), client::update);
-                    log.info("Visited {} links: \n{}", name, visited);
-                });
-            }
-        } finally {
-            threadPool.shutdown();
-        }
-    }
-
-    private GroceryDadConfig.Scraper scraperConfig(String name) {
-        if (!config.scrapers.containsKey(name)) {
-            throw new IllegalArgumentException(format("Missing %s config", name));
-        }
-        return config.scrapers.get(name);
+        scraper.scrap(client::update);
     }
 
     public static void main(String... args) {
         System.setProperty(CHROME_DRIVER_SILENT_OUTPUT_PROPERTY, "true");
         SpringApplication.run(GroceryDad.class, args);
+        System.exit(0);
     }
 }
